Only add a QrCode if the library is present

+++ fannie/ordering/SpecialOrderTags.php
@@ -157,19 +157,21 @@ class SpecialOrderTags extends FannieRESTfulPage
             $pdf = $signage->drawBarcode($upc, $pdf, $posX+10, $posY+95, array('height'=>14,'fontsize'=>8));
 
             $reorder_url = 'http://wholefoods.coop/reorder/' . $oid . '-' . $tid;
-            $qrImg = tempnam(sys_get_temp_dir(), 'qrc') . '.png';
-            $qrCode = new QrCode();
-            $qrCode->setText($reorder_url)
-                ->setSize(60)
-                ->setPadding(2)
-                ->setErrorCorrection('high')
-                ->setForegroundColor(array('r' => 0, 'g' => 0, 'b' => 0, 'a' => 0))
-                ->setBackgroundColor(array('r' => 255, 'g' => 255, 'b' => 255, 'a' => 0))
-                ->setLabelFontSize(6)
-                ->render($qrImg);
+            if (class_exists('Endroid\\QrCode\\QrCode')) {
+                $qrImg = tempnam(sys_get_temp_dir(), 'qrc') . '.png';
+                $qrCode = new QrCode();
+                $qrCode->setText($reorder_url)
+                    ->setSize(60)
+                    ->setPadding(2)
+                    ->setErrorCorrection('high')
+                    ->setForegroundColor(array('r' => 0, 'g' => 0, 'b' => 0, 'a' => 0))
+                    ->setBackgroundColor(array('r' => 255, 'g' => 255, 'b' => 255, 'a' => 0))
+                    ->setLabelFontSize(6)
+                    ->render($qrImg);
 
-            $pdf->Image($qrImg, $posX+60, $posY+93);
-            unlink($qrImg);
+                $pdf->Image($qrImg, $posX+60, $posY+93);
+                unlink($qrImg);
+            }
 
             $pdf->SetXY($posX,$posY+115);
             $pdf->Cell(90,10,"Re-Order: $reorder_url", 0, 0, 'C');
